requirements.txt
# GUI框架
PyQt5>=5.15.0                 # Qt界面库
PyQt5-sip>=12.11.0           # PyQt5的Python绑定
qtawesome>=1.2.1             # 图标库

# 计算机视觉
opencv-python>=4.5.0         # OpenCV图像处理库
ultralytics>=0.0.0           # YOLO目标检测模型
torch>=2.0.0                 # PyTorch深度学习框架
torchvision>=0.15.0          # PyTorch视觉库

# 数据处理和可视化
numpy>=1.21.0                # 数学计算库
pandas>=1.3.0                # 数据处理库
matplotlib>=3.5.0            # 数据可视化库
seaborn>=0.11.0              # 统计数据可视化

# 数据库
mysql-connector-python>=8.0.0 # MySQL数据库连接器

# 数据导出
openpyxl>=3.0.0              # Excel文件处理

